+++ integration_tests/suite/helpers/helpers/entity.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2016-2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import string
@@ -13,8 +13,8 @@ def generate_entity(**params):
     return add_entity(**params)
 
 
-def add_entity(**params):
-    response = confd.entities.post(params)
+def add_entity(wazo_tenant=None, **params):
+    response = confd.entities.post(params, wazo_tenant=wazo_tenant)
     return response.item