+++ xivo_confd/plugins/user_import/plugin.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
-# Copyright 2015-2018 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from collections import OrderedDict
 
 from xivo_dao.resources.call_permission import dao as call_permission_dao
-from xivo_dao.resources.cti_profile import dao as cti_profile_dao
 from xivo_dao.resources.endpoint_sccp import dao as sccp_dao
 from xivo_dao.resources.endpoint_sip import dao as sip_dao
 from xivo_dao.resources.entity import dao as entity_dao
@@ -22,6 +21,7 @@ from wazo_provd_client import Client as ProvdClient
 
 from xivo_confd.database import user_export as user_export_dao
 from xivo_confd.plugins.call_permission.service import build_service as build_call_permission_service
+from xivo_confd.plugins.context.service import build_service as build_context_service
 from xivo_confd.plugins.endpoint_sccp.service import build_service as build_sccp_service
 from xivo_confd.plugins.endpoint_sip.service import build_service as build_sip_service
 from xivo_confd.plugins.entity.service import build_service as build_entity_service
@@ -33,7 +33,6 @@ from xivo_confd.plugins.line_endpoint.service import build_service as build_le_s
 from xivo_confd.plugins.line_extension.service import build_service as build_line_extension_service
 from xivo_confd.plugins.user.service import build_service as build_user_service
 from xivo_confd.plugins.user_call_permission.service import build_service as build_user_call_permission_service
-from xivo_confd.plugins.user_cti_profile.service import build_service as build_user_cti_profile_service
 from xivo_confd.plugins.user_entity.service import build_service as build_user_entity_service
 from xivo_confd.plugins.user_line.service import build_service as build_ul_service
 from xivo_confd.plugins.user_voicemail.service import build_service as build_uv_service
@@ -41,7 +40,6 @@ from xivo_confd.plugins.voicemail.service import build_service as build_voicemai
 
 from .associators import (
     CallPermissionAssociator,
-    CtiProfileAssociator,
     EntityAssociator,
     ExtensionAssociator,
     IncallAssociator,
@@ -53,7 +51,7 @@ from .associators import (
 )
 from .creators import (
     CallPermissionCreator,
-    CtiProfileCreator,
+    ContextCreator,
     EntityCreator,
     ExtensionCreator,
     IncallCreator,
@@ -86,7 +84,6 @@ class Plugin(object):
         wazo_user_service = build_wazo_user_service()
         entity_service = build_entity_service()
         user_voicemail_service = build_uv_service()
-        user_cti_profile_service = build_user_cti_profile_service()
         voicemail_service = build_voicemail_service()
         line_service = build_line_service(provd_client)
         sip_service = build_sip_service(provd_client)
@@ -101,6 +98,7 @@ class Plugin(object):
         user_call_permission_service = build_user_call_permission_service()
         incall_service = build_incall_service()
         incall_extension_service = build_incall_extension_service()
+        context_service = build_context_service()
 
         creators = {
             'user': UserCreator(user_service),
@@ -113,8 +111,8 @@ class Plugin(object):
             'extension': ExtensionCreator(extension_service),
             'extension_incall': ExtensionCreator(extension_service),
             'incall': IncallCreator(incall_service),
-            'cti_profile': CtiProfileCreator(cti_profile_dao),
             'call_permissions': CallPermissionCreator(call_permission_service),
+            'context': ContextCreator(context_service),
         }
 
         entry_creator = EntryCreator(creators)
@@ -123,7 +121,6 @@ class Plugin(object):
             ('entity', EntityAssociator(user_entity_service)),
             ('wazo_user', WazoUserAssociator(wazo_user_service)),
             ('voicemail', VoicemailAssociator(user_voicemail_service)),
-            ('cti_profile', CtiProfileAssociator(user_cti_profile_service, cti_profile_dao)),
             ('sip', SipAssociator(line_sip_service)),
             ('sccp', SccpAssociator(line_sccp_service)),
             ('line', LineAssociator(user_line_service)),
@@ -140,7 +137,6 @@ class Plugin(object):
             entity_dao,
             voicemail_dao,
             user_voicemail_dao,
-            cti_profile_dao,
             line_dao,
             user_line_dao,
             line_extension_dao,